requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.5
aiosignal==1.3.2
aiounittest==1.3.0
attrs==25.3.0
azure-core==1.34.0
botbuilder-adapters-slack==4.9.2
botbuilder-core==4.16.2
botbuilder-integration-aiohttp==4.16.2
botbuilder-schema==4.16.2
botframework-connector==4.16.2
botframework-streaming==4.16.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cryptography==44.0.3
frozenlist==1.6.0
h11==0.16.0
idna==3.10
isodate==0.7.2
jsonpickle==1.4.2
msal==1.32.3
msrest==0.7.1
multidict==6.4.3
oauthlib==3.2.2
propcache==0.3.1
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
yarl==1.20.0
